video_backend/requirements.txt
flask
flask-cors
Flask-SQLAlchemy
flask-executor
psycopg2-binary
humps
pytube

pytorch
opencv
salesforce-lavis
happytransformer
scenedetect[opencv]
pydub